+++ benchopt/base.py
@@ -1,5 +1,4 @@
 import tempfile
-import numbers
 import warnings
 
 from abc import ABC, abstractmethod
@@ -89,19 +88,12 @@ class BaseSolver(ParametrizedNameMixin, DependenciesMixin, ABC):
         """
         self._objective = objective
         self._output = output
-        # XXX remove in version 1.4
+
         objective_dict = objective.get_objective()
-        if objective_dict is None:
-            assert hasattr(objective, "to_dict"), (
-                "Objective needs to implement `get_objective` that returns "
-                "a dictionary to be passed to `set_objective`"
-            )
-            warnings.warn(
-                "The method ``Objective.to_dict`` has been deprecated in "
-                "favor of ``Objective.get_objective``. Using it will no "
-                "longer work in version 1.4", FutureWarning
-            )
-            objective_dict = objective.to_dict()
+        assert objective_dict is not None, (
+            "Objective needs to implement `get_objective` that returns "
+            "a dictionary to be passed to `set_objective`"
+        )
 
         # Check if the objective is compatible with the solver
         skip, reason = self.skip(**objective_dict)
@@ -307,24 +299,6 @@ class BaseDataset(ParametrizedNameMixin, DependenciesMixin, ABC):
         if not hasattr(self, '_data') or self._data is None:
             self._data = self.get_data()
 
-            # XXX - Remove in version 1.3
-            if type(self._data) != dict:
-                import warnings
-                warnings.warn(
-                    "`get_data` should return a dict containing the data. "
-                    "The dimension should not be returned anymore and "
-                    "Objective should have a method `get_one_solution` "
-                    "to provide one feasible point. This will cause an "
-                    "error starting from version 1.3.",
-                    FutureWarning
-                )
-                dimension, data = self._data
-
-                # Make sure dimension is a tuple
-                if isinstance(dimension, numbers.Integral):
-                    dimension = (dimension,)
-                self._data = (dimension, data)
-
         return self._data
 
     # Reduce the pickling and hashing burden by only pickling class parameters.
@@ -383,8 +357,7 @@ class BaseObjective(ParametrizedNameMixin, DependenciesMixin):
         """
         ...
 
-    # to uncomment in version 1.4, once `to_dict` has been deprecated.
-    # @abstractmethod
+    @abstractmethod
     def get_objective(self):
         """Return the objective parameters for the solver.
 
@@ -445,10 +418,6 @@ class BaseObjective(ParametrizedNameMixin, DependenciesMixin):
         self._dataset = dataset
         data = dataset._get_data()
 
-        # XXX - Remove in version 1.3
-        if type(data) != dict:
-            self._dimension, data = data
-
         # Check if the dataset is compatible with the objective
         skip, reason = self.skip(**data)
         if skip:
@@ -494,36 +463,16 @@ class BaseObjective(ParametrizedNameMixin, DependenciesMixin):
         """
         return False, None
 
+    @abstractmethod
     def get_one_solution(self):
         """Return one solution for which the objective can be evaluated.
 
-        This method is mainly for testing purposes, to check that the return
-        computation and the return type of `Objective.compute`. It should
-        return an object that can be passed to compute.
-
-        By default, if `Dataset.get_data` returns a shape, this function will
-        return an array full of 0. It can be overriden to provide a different
-        point. When `Dataset.get_data` returns "object", this method must be
-        overwritten.
+        This method is mainly for testing purposes, to check that the method
+        `Objective.compute` can be called and that it returns a compatible
+        type for benchopt. The returned object will be passed to
+        ``Objective.compute``.
         """
-        if not hasattr(self, '_dimension'):
-            raise NotImplementedError(
-                "If solvers return objects, `Objective.get_one_solution` "
-                "should be overriden to return an object compatible with "
-                "`compute`."
-            )
-
-        # XXX - make this an abstract class in version 1.3
-        import warnings
-        warnings.warn(
-            "Objective should have a method `get_one_solution` "
-            "to provide one feasible point. This will cause an "
-            "error starting from version 1.3.",
-            FutureWarning
-        )
-
-        import numpy as np
-        return np.zeros(self._dimension)
+        pass
 
     # Reduce the pickling and hashing burden by only pickling class parameters.
     @staticmethod